apps/ai-agent/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
google-cloud-aiplatform==1.79.0
python-multipart==0.0.20
pydantic==2.10.6
